V3/requirements.txt
fastapi
uvicorn[standard]
streamlit
langchain
sqlalchemy
psycopg2-binary
chromadb-client
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
PyYAML
sentence-transformers
pypdf
python-docx
tiktoken
openai
pytest
python-multipart
google-cloud-storage
reportlab
fastapi
jose
chromadb
langchain_community
langchain-huggingface
langchain-huggingface langchain-chroma
google-api-python-client
google-auth-oauthlib
dropbox
msgraph-core
atlassian-python-api
